+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ---
+## [v0.6.4b] - 2021-04-26
+### Changed
+- Menu screen now pulses to the beat!
+### Optimised
+- Beatmap selection generator to only calculate star rating once
+### Fixed
+- Settings menu blocking gameplay [#2](https://github.com/Joshua-Usi/osw/issues/2)
+- Opening links in new tab using target="_blank"[#6](https://github.com/Joshua-Usi/osw/issues/6)
+---
 ## [0.6.3b] - 2021-04-25
 ### Changed
 - new hexagon background